Minor revisions of abstract and introduction
Mostly insubstantial clarifying edits.

A minor stylistic change is the removal of the first person ("we") in
favor of speaking for the document itself.

The most substantive change is to slightly de-emphasize Subject
encryption (no more "in particular") in favor of viewing it as one
aspect of the header protections we hope to offer.

This is due to what appeared to me to be the conensus we had at the
OpenPGP E-mail summit about header protection as the logical defense
against signature replay attacks.