Describe Legacy Display part completely
I believe this text captures our discussion about the Legacy Display
part, including both how it is generated and how it is interpreted or
elided by receiving MUAs.

In our discussions in Berlin, there was talk of changing the MIME type
of the Legacy Display part away from text/rfc822-headers to
text/plain.

I have *not* included that change in this commit because I wanted to
document current practice first, clearly.  If we decide to make that
change, i hope we can make it as a proposed change to the text
provided here.