Generate signed-message-only test vector
This is an awful lot of python to generate a reasonably small test
vector -- the test message itself is significantly smaller :)

However, it does what it needs to do and each step is sensible and
minimal.

I've included its output in the draft itself.  Hopefully we can re-use
the same framework to generate test vectors for other common forms of
Protected Header messages.