Test Vectors: add multilayer examples
Not all mailers are going to do encryption and signing at the PGP
layer.  Some will generate a multilayer cryptographic envelope.

But every mailer is going to need to accept messages with this kind of
cryptographic envelope.

Providing test vectors for multilayer envelopes should help
implementors ensure that their messages are displayed reasonably.