Test Vectors: broke out test vector generation
Generated files are easier to keep track of, if there are more of
them, than trying to keep them inline.

This assembly-based approach is a bit better, and it will help us keep
track of the different changes to the generated files, though it might
lead to some extra churn if folks aren't careful about what changes
get committed.

Unfortunately, the encrypted test vectors are not reproducible because
the encryption layer itself is non-deterministic.  So only commit
changes to those files if either:

 * the text outside of the encryption ("outside headers, etc") has
   changed, or

 * the text inside the encryption ("inner part") has changed.

This test for update should probably be automated to help people avoid
churn, but that's more complex than i'm prepared to do right now.

Signed-off-by: Daniel Kahn Gillmor <dkg@fifthhorseman.net>