Revise Summary, move to the top
The summary is an important part of the document.  Someone who reads
the summary and understands cryptographic e-mail in general should get
a sense of what's happening here without reading much more.

I've moved it to a more prominent place in the document, and i've
stripped it of language that blends together "what are the underlying
changes?" with "how should one construct such a message?" -- we want
this section to focus on the former, and leave details of the latter
to later sections.

Signed-off-by: Daniel Kahn Gillmor <dkg@fifthhorseman.net>