Composition: break out inputs from algorithm, describe current practice
I think this structure makes the algorithm more understandable, and
provides a simple contrast with the steps that legacy MUA clients are
doing.

It also makes it easier to parameterize the process.  At the moment, i
think we want the defaults to be `legacy` = True and `obscures` =
`{'Subject': '...'}`.  But in the future, some MUAs might experiment
with adjusting either variable and can report back on that experience,
all while still using the same message composition algorithm.

Signed-off-by: Daniel Kahn Gillmor <dkg@fifthhorseman.net>