Drop "Examples" section
The "Examples" section contained a text that was difficult to read, in
a non-standard format.  It was particularly unclear to me what the E:
and S: prefixes were supposed to mean, and it was distracting to see
pieces that are not actual parts of the message rendered in brackets.

Most examples also included unnecessary/distracting text like a MIME
preamble.

Furthermore, the Legacy Display parts included unobscured headers,
which is not a recommended practice.

Now that the Test Vectors include the cleartext, most of this
functionality is superceded anyway, and it is superceded by messages
that can be parsed directly in tests.

If there are flaws in the Test Vectors, we should fix those.

Signed-off-by: Daniel Kahn Gillmor <dkg@fifthhorseman.net>