Textual cleanup
I've tried to convert things from plural to singular where possible,
as that makes reading a bit easier.

I've moved multiline paragraphs into one-line-per-sentence for more
intelligible diffs.

I've identified the two workarounds for dealing with reduced metadata
for indexes as distinct choices (either/or), not a sequence of choices
(do step 1, do step 2).

I've removed some references to "We" or "the authors" to let the
document speak for itself.

Signed-off-by: Daniel Kahn Gillmor <dkg@fifthhorseman.net>