Shuffle line spacing to placate idnits
idnits 2.16.02 does line length checks based on octet count, not on
character count.

This line below:

---
   In the diagrams below, "↧" (DOWNWARDS ARROW FROM BAR, U+21A7) is used
---

was generated via xml2rfc at 72 characters, but because it includes a
non-ascii character it is more than 72 octets.

By changing the text of the paragraph, i've made the line wrap
differently so that both octet-count and char-count come in under 72
there.

This is a pretty silly workaround, but i don't have the capacity to
fix the tooling yet.

Signed-off-by: Daniel Kahn Gillmor <dkg@fifthhorseman.net>