Generate xml2rfc v3 XML
By doing the explicit translation to v3, we should be able to get the
pretty version of the HTMLizer to run on tools.ietf.org with the next
upload.

We do this by sticking with the kramdown output initially (v2 of the
rfc2xml format), and then explicitly translating it to v3 before doing
the rest of the conversions.